Remove old networking code, remove FutureNova, rename models  (#163)
* Remove unneeded files

* Create new model files

* Uncomment out interests and groups in profile view

* Uncomment out commented out code

* Address Reade's comments
