+++ Pear/AppDelegate.swift
@@ -7,7 +7,6 @@
 //
 
 import AppDevAnnouncements
-import FutureNova
 import GoogleSignIn
 import UIKit
 
@@ -26,9 +25,6 @@ class AppDelegate: UIResponder, UIApplicationDelegate {
             commonPath: Keys.announcementsCommonPath,
             announcementPath: Keys.announcementsPath
         )
-
-        // Setup networking
-        Endpoint.setupEndpointConfig()
         
         return true
     }
