--- Pear/Controller/Group.swift
@@ -1,16 +1,0 @@
-//
-//  Group.swift
-//  CoffeeChat
-//
-//  Created by Phillip OReggio on 2/11/20.
-//  Copyright © 2020 cuappdev. All rights reserved.
-//
-
-import Foundation
-
-struct Group {
-
-    let name: String
-    let imageName: String
-
-}
