--- Pear/Models/Interest.swift
@@ -1,17 +1,0 @@
-//
-//  Interest.swift
-//  CoffeeChat
-//
-//  Created by Phillip OReggio on 2/4/20.
-//  Copyright © 2020 cuappdev. All rights reserved.
-//
-
-import Foundation
-
-struct Interest {
-
-    let name: String
-    let categories: [String]?
-    let imageName: String
-
-}
