--- Pear/Models/SubUser.swift
@@ -1,25 +1,0 @@
-//
-//  SubUser.swift
-//  CoffeeChat
-//
-//  Created by Lucy Xu on 5/7/20.
-//  Copyright © 2020 cuappdev. All rights reserved.
-//
-
-import Foundation
-
-struct SubUser: Codable {
-
-    let firstName: String
-    let googleID: String
-    let graduationYear: String
-    let hometown: String
-    let lastName: String
-    let major: String
-    let netID: String
-    let profilePictureURL: String?
-    let pronouns: String
-    let facebook: String?
-    let instagram: String?
-
-}
