+++ Pear/Models/User.swift
@@ -5,7 +5,6 @@
 //  Created by Lucy Xu on 5/7/20.
 //  Copyright © 2020 cuappdev. All rights reserved.
 //
-
 import Foundation
 
 struct User: Codable, Equatable {
