+++ Pear/Models/UserV2.swift
@@ -23,8 +23,8 @@ struct UserV2: Codable {
     let talkingPoints: [String]?
     let availability: [String]?
     let locations: [LocationV2]?
-    let interests: [InterestV2]
-    let groups: [GroupV2]
+    let interests: [Interest]
+    let groups: [Group]
     let hasOnboarded: Bool?
     let pendingFeedback: Bool
     let currentMatch: MatchV2?
@@ -43,12 +43,7 @@ struct MatchV2: Codable {
     let meetingTime: String?
 }
 
-protocol Topic {
-    var name: String { get set }
-    var id: Int { get set }
-    var imgUrl: String { get set}
-}
-
+// TODO: MatchedUser and CommunityUser can be combined
 struct MatchedUser: Codable {
     let id: Int
     let netId: String
@@ -58,9 +53,9 @@ struct MatchedUser: Codable {
     let majors: [MajorV2]
     let profilePicUrl: String?
     let graduationYear: String
-    let pronouns: String
-    let interests: [InterestV2]
-    let groups: [GroupV2]
+    let pronouns: String?
+    let interests: [Interest]
+    let groups: [Group]
 }
 
 struct CommunityUser: Codable {
@@ -68,39 +63,18 @@ struct CommunityUser: Codable {
     let netId: String
     let firstName: String
     let lastName: String
+    let hometown: String
     let majors: [MajorV2]
     let profilePicUrl: String?
-    let hometown: String
     let graduationYear: String
-    let interests: [InterestV2]
-    let groups: [GroupV2]
+    let interests: [Interest]
+    let groups: [Group]
 }
 
 struct LocationV2: Codable {
     let id: String
     let name: String
     let area: String
-}
-
-struct InterestV2: Topic, Codable {
-    var id: Int
-    var name: String
-    let subtitle: String
-    var imgUrl: String
-}
-
-struct GroupV2: Topic, Codable {
-    var id: Int
-    var name: String
-    var imgUrl: String
-}
-
-struct TalkingPointV2: Codable {
-    let type: String
-    let id: String
-    let name: String
-    let subtitle: String?
-    let imgUrl: String
 }
 
 // todo - fix naming
