+++ Pear/Onboarding/InterestsViewController.swift
@@ -22,8 +22,8 @@ class InterestsViewController: UIViewController {
 
     // MARK: - Data
     private weak var delegate: OnboardingPageDelegate?
-    private var interests: [InterestV2] = []
-    private var selectedInterests: [InterestV2] = []
+    private var interests: [Interest] = []
+    private var selectedInterests: [Interest] = []
 
     init(delegate: OnboardingPageDelegate) {
         self.delegate = delegate
@@ -78,7 +78,7 @@ class InterestsViewController: UIViewController {
 
     @objc func nextButtonPressed() {
         let selectedInterestsIds = selectedInterests.map { $0.id }
-        Networking2.updateInterests(interests: selectedInterestsIds) { [weak self] (success) in
+        NetworkManager.updateInterests(interests: selectedInterestsIds) { [weak self] (success) in
             guard let self = self else { return }
             DispatchQueue.main.async {
                 if success {
@@ -123,7 +123,7 @@ class InterestsViewController: UIViewController {
 
     private func getAllInterests() {
 
-        Networking2.getAllInterests { [weak self] interests in
+        NetworkManager.getAllInterests { [weak self] interests in
             guard let self = self else { return }
             DispatchQueue.main.async {
                 self.interests = interests
@@ -135,7 +135,7 @@ class InterestsViewController: UIViewController {
 
     private func getUserInterests() {
 
-        Networking2.getMe { [weak self] user in
+        NetworkManager.getMe { [weak self] user in
             guard let self = self else { return }
             DispatchQueue.main.async {
                 self.selectedInterests = user.interests
